% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{as.sql}
\alias{as.sql}
\title{Convert to sql (deprecated)}
\usage{
as.sql(x, con)
}
\arguments{
\item{x}{Object to coerce}

\item{con}{Needed when \code{x} is directly supplied from the user so that
schema specifications can be quoted using the correct identifiers.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\keyword{internal}
