% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-.R
\name{backend-ansi}
\alias{dialect_ansi}
\alias{simulate_dbi}
\title{ANSI SQL}
\usage{
dialect_ansi()

simulate_dbi(class = character(), ...)
}
\arguments{
\item{class, ...}{No longer used.}
}
\description{
This is the base dialect for ANSI compliant SQL, forming the foundation
of all othe dialects. See \code{vignette("translation-function")} for a list
of functions that are translated.

Use \code{dialect_ansi()} with \code{lazy_frame()} to see simulated SQL without
connecting to a live database.
}
\examples{
library(dplyr, warn.conflicts = FALSE)

lf <- lazy_frame(a = TRUE, b = 1, c = 2, d = "z", con = dialect_ansi())
lf |> transmute(x = mean(b, na.rm = TRUE))
lf |> transmute(x = log(b), y = log(b, base = 2))
}
