% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-hana.R
\name{backend-hana}
\alias{dialect_hana}
\alias{simulate_hana}
\title{SAP HANA backend}
\usage{
dialect_hana()

simulate_hana()
}
\description{
This backend supports SAP HANA databases, typically accessed via
\code{HDBConnection} created by \code{\link[DBI:dbConnect]{DBI::dbConnect()}}. Use \code{dialect_hana()} with
\code{lazy_frame()} to see simulated SQL without connecting to a live database.

Key differences for this backend are:
\itemize{
\item Temporary tables get \verb{#} prefix and use \verb{LOCAL TEMPORARY COLUMN}.
\item No table analysis performed in \code{\link[dplyr:copy_to]{dplyr::copy_to()}}.
\item \code{paste()} uses \code{||}
\item Note that you can't create new boolean columns from logical expressions;
you need to wrap with explicit \code{ifelse}: \code{ifelse(x > y, TRUE, FALSE)}.
}

See \code{vignette("translation-function")} and \code{vignette("translation-verb")} for
details of overall translation technology.
}
\examples{
library(dplyr, warn.conflicts = FALSE)

lf <- lazy_frame(a = TRUE, b = 1, c = 2, d = "z", con = dialect_hana())
lf |> transmute(x = paste0(d, " times"))
}
