% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-hive.R
\name{backend-hive}
\alias{dialect_hive}
\alias{simulate_hive}
\title{Hive backend}
\usage{
dialect_hive()

simulate_hive()
}
\description{
This backend supports Apache Hive, typically accessed via odbc. Use
\code{dialect_hive()} with \code{lazy_frame()} to see simulated SQL without connecting
to a live database.

Key differences for this backend are a scattering of custom translations
provided by users.

See \code{vignette("translation-function")} and \code{vignette("translation-verb")} for
details of overall translation technology.
}
\examples{
library(dplyr, warn.conflicts = FALSE)

lf <- lazy_frame(a = TRUE, b = 1, d = 2, c = "z", con = dialect_hive())
lf |> transmute(x = cot(b))
lf |> transmute(x = bitwShiftL(c, 1L))
lf |> transmute(x = str_replace_all(c, "a", "b"))

lf |> summarise(x = median(d, na.rm = TRUE))
lf |> summarise(x = var(c, na.rm = TRUE))
}
