% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-impala.R
\name{backend-impala}
\alias{dialect_impala}
\alias{simulate_impala}
\title{Impala backend}
\usage{
dialect_impala()

simulate_impala()
}
\description{
This backend supports Apache Impala, typically accessed via odbc. Use
\code{dialect_impala()} with \code{lazy_frame()} to see simulated SQL without
connecting to a live database.

Key differences for this backend are a scattering of custom translations
provided by users, mostly focussed on bitwise operations.

See \code{vignette("translation-function")} and \code{vignette("translation-verb")} for
details of overall translation technology.
}
\examples{
library(dplyr, warn.conflicts = FALSE)

lf <- lazy_frame(a = TRUE, b = 1, c = 2, d = "z", con = dialect_impala())
lf |> transmute(X = bitwNot(bitwOr(b, c)))
}
