% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-odbc.R
\name{backend-odbc}
\alias{simulate_odbc}
\title{Backend: ODBC}
\usage{
simulate_odbc()
}
\description{
See \code{vignette("translation-function")} and \code{vignette("translation-verb")} for
details of overall translation technology. Key differences for this backend
are minor translations for common data types.

Use \code{simulate_odbc()} with \code{lazy_frame()} to see simulated SQL without
converting to live access database.
}
\examples{
library(dplyr, warn.conflicts = FALSE)

lf <- lazy_frame(a = TRUE, b = 1, d = 2, c = "z", con = simulate_odbc())
lf |> transmute(x = as.numeric(b))
lf |> transmute(x = as.integer(b))
lf |> transmute(x = as.character(b))
}
