% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-redshift.R
\name{backend-redshift}
\alias{dialect_redshift}
\alias{simulate_redshift}
\title{Redshift backend}
\usage{
dialect_redshift()

simulate_redshift()
}
\description{
Base translations come from \link[=dialect_postgres]{PostgreSQL backend}. There
are generally few differences, apart from string manipulation.

Use \code{dialect_redshift()} with \code{lazy_frame()} to see simulated SQL without
connecting to a live database.
}
\examples{
library(dplyr, warn.conflicts = FALSE)

lf <- lazy_frame(a = TRUE, b = 1, c = 2, d = "z", con = dialect_redshift())
lf |> transmute(x = paste(c, " times"))
lf |> transmute(x = substr(c, 2, 3))
lf |> transmute(x = str_replace_all(c, "a", "z"))
}
