% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-teradata.R
\name{backend-teradata}
\alias{simulate_teradata}
\alias{dialect_teradata}
\title{Backend: Teradata}
\usage{
simulate_teradata()

dialect_teradata()
}
\description{
See \code{vignette("translation-function")} and \code{vignette("translation-verb")} for
details of overall translation technology. Key differences for this backend
are:
\itemize{
\item Uses \code{TOP} instead of \code{LIMIT}
\item Selection of user supplied translations
}

Use \code{simulate_teradata()} with \code{lazy_frame()} to see simulated SQL without
converting to live access database.
}
\examples{
library(dplyr, warn.conflicts = FALSE)

lf <- lazy_frame(a = TRUE, b = 1, c = 2, d = "z", con = simulate_teradata())
lf |> head()
}
