% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-sql.R
\name{build_sql}
\alias{build_sql}
\title{Build a SQL string.}
\usage{
build_sql(..., .env = parent.frame(), con = sql_current_con())
}
\arguments{
\item{...}{input to convert to SQL. Use \code{\link[=sql]{sql()}} to preserve
user input as is (dangerous), and \code{\link[=ident]{ident()}} to label user
input as sql identifiers (safe)}

\item{.env}{the environment in which to evaluate the arguments. Should not
be needed in typical use.}

\item{con}{database connection; used to select correct quoting characters.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{build_sql()} is deprecated in favor of \code{glue_sql2()}.

This is a convenience function that should prevent sql injection attacks
(which in the context of dplyr are most likely to be accidental not
deliberate) by automatically escaping all expressions in the input, while
treating bare strings as sql. This is unlikely to prevent any serious
attack, but should make it unlikely that you produce invalid sql.

This function should be used only when generating \code{SELECT} clauses,
other high level queries, or for other syntax that has no R equivalent.
For individual function translations, prefer \code{\link[=sql_expr]{sql_expr()}}.
}
\examples{
con <- simulate_dbi()

# Old:
build_sql("SELECT * FROM ", ident("table"), con = con)
# New:
glue_sql2(con, "SELECT * FROM {.tbl 'table'}")

# Old:
name <- "Robert"
build_sql("INSERT INTO students (name) VALUES (", name, ")", con = con)
# New:
glue_sql2(con, "INSERT INTO students (name) VALUES ({.val name})")

# Old:
cols <- c("x", "y")
build_sql("SELECT ", ident(cols), " FROM table", con = con)
# New:
glue_sql2(con, "SELECT {.col cols*} FROM {.tbl 'table'}")
}
\keyword{internal}
