% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyr-pivot.R, R/verb-compute.R,
%   R/verb-copy-to.R
\name{db_pivot_wider}
\alias{db_pivot_wider}
\alias{db_compute}
\alias{db_collect}
\alias{db_sql_render}
\alias{db_copy_to}
\title{More db generics}
\usage{
db_pivot_wider(data, id_cols = NULL, names_from = name,
  names_prefix = "", names_sep = NULL, names_repair = NULL,
  values_from = value, values_fill = NULL, values_fn = NULL,
  spec = NULL)

db_compute(con, table, sql, temporary = TRUE, unique_indexes = list(),
  indexes = list(), analyze = TRUE, ...)

db_collect(con, sql, n = -1, warn_incomplete = TRUE, ...)

db_sql_render(con, sql, ...)

db_copy_to(con, table, values, overwrite = FALSE, types = NULL,
  temporary = TRUE, unique_indexes = NULL, indexes = NULL,
  analyze = TRUE, ...)
}
\description{
These are new, so not included in dplyr for backward compatibility
purposes.
}
\keyword{internal}
