% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbplyr.R
\docType{package}
\name{dbplyr-package}
\alias{dbplyr}
\alias{dbplyr-package}
\title{dbplyr: A 'dplyr' Backend for Databases}
\description{
A dplyr backend for databases that allows you to work with
remote database tables as if they are in-memory data frames. Basic features
works with any database that has a DBI backend; more advanced features
require SQL translation to be provided by the package author.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tidyverse/dbplyr}
  \item Report bugs at \url{https://github.com/tidyverse/dplyr/issues}
}

}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
}

}
\keyword{internal}
