% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verb-fill.R
\name{dbplyr_fill}
\alias{dbplyr_fill}
\title{Fill in missing values with previous or next value}
\usage{
dbplyr_fill(.data, ..., order_by, .direction = c("down", "up"))
}
\arguments{
\item{.data}{A lazy data frame backed by a database query.}

\item{...}{<\code{\link[=tidyr_tidy_select]{tidy-select}}> Columns to fill.}

\item{order_by}{Columns to order by before filling the values.}

\item{.direction}{Direction in which to fill missing values. Currently
either "down" (the default) or "up".}
}
\description{
Fill in missing values with previous or next value
}
\examples{
squirrels <- tibble::tribble(
  ~group,    ~name,     ~role,     ~n_squirrels, ~ n_squirrels2,
  1,      "Sam",    "Observer",   NA,                 1,
  1,     "Mara", "Scorekeeper",    8,                NA,
  1,    "Jesse",    "Observer",   NA,                NA,
  1,      "Tom",    "Observer",   NA,                 4,
  2,     "Mike",    "Observer",   NA,                NA,
  2,  "Rachael",    "Observer",   NA,                 6,
  2,  "Sydekea", "Scorekeeper",   14,                NA,
  2, "Gabriela",    "Observer",   NA,                NA,
  3,  "Derrick",    "Observer",   NA,                NA,
  3,     "Kara", "Scorekeeper",    9,                 10,
  3,    "Emily",    "Observer",   NA,                NA,
  3, "Danielle",    "Observer",   NA,                NA
)
squirrels$id <- 1:12

dbplyr_fill(
  tbl_memdb(squirrels),
  n_squirrels,
  n_squirrels2,
  order_by = c(id)
)
}
