% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyeval.R
\docType{data}
\name{.sql}
\alias{.sql}
\title{Flag SQL function usage}
\usage{
.sql
}
\description{
Use \code{.sql$foo(x, y)} to make it clear that you're calling the SQL
\code{foo()} function, not the R \code{foo()} function. This also makes it easier to
reduce \verb{R CMD check} notes in packages; just import \code{.sql} from dbplyr with
e.g. \verb{@importFrom dbplyr .sql}.

Note that \code{.sql} itself does nothing and is just \code{NULL}; it is automatically
removed when dbplyr translates your R code to SQL.
}
\examples{
library(dplyr, warn.conflicts = FALSE)

db <- lazy_frame(x = 1, y = 2)
db |> mutate(z = .sql$CUMULATIVE_SUM(x, 1))
}
\keyword{datasets}
