% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward-compatibility.R
\name{escape_ansi}
\alias{escape_ansi}
\title{Escape ANSI characters}
\usage{
escape_ansi(x, parens = NA, collapse = "")
}
\arguments{
\item{x}{Object to escape.}

\item{parens, collapse}{Controls parens and collapsing. Passed on to
\code{\link[=escape]{escape()}}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Use \code{\link[=escape]{escape()}} instead.
}
\keyword{internal}
