% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verb-filter.R
\name{filter.tbl_lazy}
\alias{filter.tbl_lazy}
\title{Subset rows using column values}
\usage{
\method{filter}{tbl_lazy}(.data, ..., .preserve = FALSE)
}
\arguments{
\item{.data}{A lazy data frame backed by a database query.}

\item{...}{<\code{\link[dplyr:dplyr_data_masking]{data-masking}}> Variables, or functions or
variables. Use \code{\link[dplyr:desc]{desc()}} to sort a variable in descending order.}

\item{.preserve}{Not supported by this method.}
}
\value{
Another \code{tbl_lazy}. Use \code{\link[=show_query]{show_query()}} to see the generated
query, and use \code{\link[=collect.tbl_sql]{collect()}} to execute the query
and return data to R.
}
\description{
This is a method for the dplyr \code{\link[=filter]{filter()}} generic. It generates the
\code{WHERE} clause of the SQL query.
}
\examples{
library(dplyr, warn.conflicts = FALSE)

db <- memdb_frame(x = c(2, NA, 5, NA, 10), y = 1:5)
db \%>\% filter(x < 5) \%>\% show_query()
db \%>\% filter(is.na(x)) \%>\% show_query()
}
