% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query-join.R, R/query-select.R,
%   R/query-semi-join.R, R/query-set-op.R, R/query.R
\name{sql_join}
\alias{sql_join}
\alias{sql_select}
\alias{sql_subquery}
\alias{sql_semi_join}
\alias{sql_set_op}
\alias{generic-query}
\title{Query generics}
\usage{
sql_join(con, x, y, vars, type = "inner", by = NULL, ...)

sql_select(
  con,
  select,
  from,
  where = NULL,
  group_by = NULL,
  having = NULL,
  order_by = NULL,
  limit = NULL,
  distinct = FALSE,
  ...
)

sql_subquery(con, from, name = random_table_name(), ...)

sql_semi_join(con, x, y, anti = FALSE, by = NULL, ...)

sql_set_op(con, x, y, method)
}
\arguments{
\item{con}{A database connection.}
}
\description{
These generics allow backends to customise the SQL generated for various
types of query. The default \code{DBIConnection} methods generate ANSI 92
compliant SQL.
}
\keyword{internal}
