% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ident.R
\name{ident}
\alias{ident}
\alias{is.ident}
\alias{as.ident}
\title{Flag a character vector as SQL identifiers}
\usage{
ident(...)

is.ident(x)

as.ident(x)
}
\arguments{
\item{...}{A character vector, or name-value pairs}

\item{x}{An object}
}
\description{
This ensures that during SQL generation, the values will be quoted as
identifiers, not as strings. \code{as.ident()} coerces to an identifier,
assuming that escaped SQL already represents an identifier.
}
\examples{
# SQL92 quotes strings with '
escape("x")

# And identifiers with "
escape(ident("x"))
}
