% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-superseded.R
\name{ident_q}
\alias{ident_q}
\title{Declare a identifier as being pre-quoted.}
\usage{
ident_q(...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}

No longer needed; please use \code{\link[=sql]{sql()}} instead.
}
\keyword{internal}
