% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote.R
\name{last_sql}
\alias{last_sql}
\title{Retrieve the last SQL query generated}
\usage{
last_sql()
}
\value{
A SQL string, or \code{NULL} if no query has been generated yet.
}
\description{
This is a helper function that retrieves the most recent SQL query generated
by dbplyr, which can be useful for debugging.
}
\examples{
library(dplyr, warn.conflicts = FALSE)

df <- lazy_frame(x = 1:3)
df |> filter(x > 1)

last_sql()
}
