% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query-base.R
\name{lazy_base_query}
\alias{lazy_base_query}
\title{Create a base lazy query}
\usage{
lazy_base_query(x, vars, class = character(), ...)
}
\arguments{
\item{x}{A data source, typically a table identifier created by \code{\link[=in_schema]{in_schema()}}
or \code{\link[=in_catalog]{in_catalog()}}, or a literal \code{\link[=sql]{sql()}} string.}

\item{vars}{A character vector of column names.}

\item{class}{A character vector of additional subclasses to add. The
resulting object will have class \code{c("lazy_base_{class}_query", "lazy_base_query", "lazy_query")}.}

\item{...}{Additional arguments passed to \code{\link[=lazy_query]{lazy_query()}}.}
}
\value{
A lazy query object.
}
\description{
\code{lazy_base_query()} is a constructor for base lazy query objects. A base
lazy query represents the root of a lazy query tree, i.e. a database table or
query.
}
\keyword{internal}
