% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy-ops.R, R/verb-filter.R, R/verb-select.R,
%   R/verb-summarise.R
\name{lazy_ops}
\alias{lazy_ops}
\alias{lazy_query_base}
\alias{op_grps}
\alias{op_vars}
\alias{op_sort}
\alias{op_frame}
\alias{add_filter}
\alias{add_select}
\alias{add_summarise}
\title{Lazy operations}
\usage{
lazy_query_base(x, vars, class = character())

op_grps(op)

op_vars(op)

op_sort(op)

op_frame(op)

add_filter(.data, dots)

add_select(.data, vars, op = c("select", "mutate"))

add_summarise(.data, dots, .groups, env_caller)
}
\description{
This set of S3 classes describe the action of dplyr verbs. These are
currently used for SQL sources to separate the description of operations
in R from their computation in SQL. This API is very new so is likely
to evolve in the future.
}
\details{
\code{op_vars()} and \code{op_grps()} compute the variables and groups from
a sequence of lazy operations. \code{op_sort()} and \code{op_frame()} tracks the
order and frame for use in window functions.
}
\keyword{internal}
