% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{named_commas}
\alias{named_commas}
\title{Provides comma-separated string out of the parameters}
\usage{
named_commas(x)
}
\arguments{
\item{...}{Arguments to be constructed into the string}
}
\description{
Provides comma-separated string out of the parameters
}
\keyword{internal}
