% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-generics.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{union}
\alias{reexports}
\alias{intersect}
\alias{reexports}
\alias{setdiff}
\alias{reexports}
\alias{filter}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{union}}, \code{\link[dplyr]{intersect}}, \code{\link[dplyr]{setdiff}}, \code{\link[dplyr]{filter}}}
}}

