% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rows.R
\name{rows_update.tbl_lazy}
\alias{rows_update.tbl_lazy}
\alias{sql_rows_update}
\alias{sql_returning_cols}
\title{sql_returning_cols}
\usage{
\method{rows_update}{tbl_lazy}(
  x,
  y,
  by = NULL,
  ...,
  copy = FALSE,
  in_place = NULL,
  check = NULL,
  returning = NULL
)

sql_rows_update(x, y, by, ..., returning_cols = NULL)

sql_returning_cols(x, returning_cols, ...)
}
\description{
\code{sql_returning_cols()} and \code{sql_output_cols()} construct the SQL
required to support the \code{returning} argument.
Two methods are required, because the syntax for SQL Server
(and some other databases) is vastly different from Postgres and other
more standardized DBs.
}
