% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.r
\name{show_query}
\alias{show_query}
\title{Show query}
\usage{
show_query(x)
}
\arguments{
\item{x}{A \link{tbl_dbi}}
}
\description{
Show query
}
\examples{
library(dplyr)

lahman_s <- lahman_sqlite()
batting <- tbl(lahman_s, "Batting")
batting \%>\% show_query()
}
