% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{show_query.tbl_lazy}
\alias{show_query.tbl_lazy}
\alias{explain.tbl_sql}
\alias{show_sql}
\title{Show the query backing a tbl}
\usage{
\method{show_query}{tbl_lazy}(x, ..., cte = FALSE, sql_options = NULL)

\method{explain}{tbl_sql}(x, ...)

show_sql(x, ...)
}
\arguments{
\item{x}{An object to explain}

\item{...}{For \code{explain()}, further arguments to \code{\link[=remote_query_plan]{remote_query_plan()}}.
For \code{show_sql()}, further arguments to other methods. For \code{show_query()},
ignored.}

\item{cte}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Use the \code{sql_options} argument instead.}

\item{sql_options}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
SQL rendering options generated by \code{\link[=sql_options]{sql_options()}}.}
}
\description{
These functions all break the abstraction layer and directly display
the underlying query that will be dispatched to the SQL engine
backing a connection.
}
\section{Functions}{
\itemize{
\item \code{explain(tbl_sql)}: Combine the \code{show_query()} and \code{remote_query_plan()}
summaries of \code{x}.

}}
