% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_dbi}
\alias{simulate_dbi}
\alias{simulate_access}
\alias{simulate_hive}
\alias{simulate_mysql}
\alias{simulate_impala}
\alias{simulate_mssql}
\alias{simulate_odbc}
\alias{simulate_oracle}
\alias{simulate_postgres}
\alias{simulate_sqlite}
\alias{simulate_teradata}
\title{Simulate database connections}
\usage{
simulate_dbi(class = character())

simulate_access()

simulate_hive()

simulate_mysql()

simulate_impala()

simulate_mssql()

simulate_odbc()

simulate_oracle()

simulate_postgres()

simulate_sqlite()

simulate_teradata()
}
\description{
These functions generate S3 objects that have been designed to simulate
the action of a database connection, without actually having the database
available. Obviously, this simulation can only be incomplete, but most
importantly it allows us to simulate SQL generation for any database without
actually connecting to it.
}
\details{
Simulated SQL always quotes identifies with \code{`x`}, and strings with
\code{'x'}.
}
\keyword{internal}
