% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{sql}
\alias{sql}
\alias{is.sql}
\title{Literal SQL escaping}
\usage{
sql(...)

is.sql(x)
}
\arguments{
\item{...}{Character vectors that will be combined into a single SQL vector.}

\item{x}{Object to check if it is an sql object.}
}
\description{
Use \code{sql()} to declare that a string is literal SQL and should be used
as is, without quoting.
}
\examples{
library(dplyr, warn.conflicts = FALSE)
# sql() just adds a class
sql("x + 1")
is.sql(sql("x + 1"))

# You can use it when you need to insert some literal SQL in a query
db <- memdb_frame(x = 1:3)
db |> mutate(y = sql("CAST(x as VARCHAR)"))
}
