% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate-sql-window.r
\name{sql_current_context}
\alias{sql_current_context}
\title{Function that retrieves a list of arguments to handle special cases.}
\usage{
sql_current_context()
}
\description{
Function that retrieves a list of arguments to handle special cases.
}
\examples{
sql_current_context()clause
}
