% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escape.R
\name{sql_escape}
\alias{sql_escape}
\alias{sql_escape_logical}
\alias{sql_escape_string}
\alias{sql_escape_ident}
\title{SQL escaping generics}
\usage{
sql_escape_logical(con, x)

sql_escape_string(con, x)

sql_escape_ident(con, x)
}
\description{
These generics provide ways for backends to override escaping on a
variable-by-variable basis, if
These are new, so not included in dplyr for backward compatibility
purposes.
}
\keyword{internal}
