% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-generic.R
\name{sql_escape_logical}
\alias{sql_escape_logical}
\title{More SQL generics}
\usage{
sql_escape_logical(con, x)
}
\description{
These are new, so not included in dplyr for backward compatibility
purposes.
}
\keyword{internal}
