% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escape.R
\name{sql_quote}
\alias{sql_quote}
\title{Helper function for quoting sql elements.}
\usage{
sql_quote(x, quote)
}
\arguments{
\item{x}{Character vector to escape.}

\item{quote}{Quote character. Either a length 1 character vector for
symmetric quotes (e.g., \code{"'"} or \code{'"'}), or a length 2 character vector
for asymmetric quotes (e.g., \code{c("[", "]")}).}
}
\description{
If the quote character is present in the string, it will be doubled.
\code{NA}s will be replaced with NULL.
}
\examples{
sql_quote("abc", "'")
sql_quote("I've had a good day", "'")
sql_quote(c("abc", NA), "'")

sql_quote(c("abc", NA), c("[", "]"))
}
\keyword{internal}
