% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escape.R
\name{sql_raw_to_hex}
\alias{sql_raw_to_hex}
\title{Helper function to convert raw vector to hex string}
\usage{
sql_raw_to_hex(x)
}
\arguments{
\item{x}{Raw vector}
}
\description{
Helper function to convert raw vector to hex string
}
\examples{
sql_raw_to_hex(charToRaw("abc"))
}
\keyword{internal}
