% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_dbi.R
\name{src_dbi}
\alias{src_dbi}
\title{Database src}
\usage{
src_dbi(con, auto_disconnect = FALSE)
}
\arguments{
\item{con}{An object that inherits from \link[DBI:DBIConnection-class]{DBI::DBIConnection},
typically generated by \link[DBI:dbConnect]{DBI::dbConnect}}

\item{auto_disconnect}{Should the connection be automatically closed when
the src is deleted? Set to \code{TRUE} if you initialize the connection
the call to \code{src_dbi()}. Pass \code{NA} to auto-disconnect but print a message
when this happens.}
}
\value{
An S3 object with class \code{src_dbi}, \code{src_sql}, \code{src}.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#superseded'><img src='figures/lifecycle-superseded.svg' alt='Superseded lifecycle'></a>}}{\strong{Superseded}}

Since can generate a \code{tbl()} directly from a DBI connection we no longer
recommend using \code{src_dbi()}.
}
\keyword{internal}
