% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memdb.R
\name{src_memdb}
\alias{src_memdb}
\alias{tbl_memdb}
\title{Deprecated}
\usage{
src_memdb()

tbl_memdb(df, name = deparse(substitute(df)))
}
\arguments{
\item{df}{Data frame to copy.}

\item{name}{Name of table in database.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\itemize{
\item \code{src_memdb()} is deprecated; use \code{memdb()} instead.
\item \code{tbl_memdb(df)} is deprecated; use \code{copy_to(memdb(), df)} instead.
}
}
\keyword{internal}
