% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward-compatibility.R
\name{src_sql}
\alias{src_sql}
\title{Create a "sql src" object}
\usage{
src_sql(subclass, con, ...)
}
\arguments{
\item{subclass}{Name of subclass.}

\item{con}{The connection object.}

\item{...}{Other arguments passed on to individual methods.}
}
\description{
Deprecated: please use directly use a \code{DBIConnection} object instead.
}
\keyword{internal}
