% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl-sql.R
\name{tbl_sql}
\alias{tbl_sql}
\title{Create an SQL tbl (abstract)}
\usage{
tbl_sql(subclass, src, from, ..., vars = NULL)
}
\arguments{
\item{subclass}{name of subclass}

\item{...}{needed for agreement with generic. Not otherwise used.}

\item{vars}{DEPRECATED}
}
\description{
Generally, you should no longer need to provide a custom \code{tbl()}
method you you can default \code{tbl.DBIConnect} method.
}
\keyword{internal}
