% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-frame.R
\name{testing}
\alias{testing}
\title{Infrastructure for testing dplyr}
\description{
Register testing sources, then use \code{test_load()} to load an existing
data frame into each source. To create a new table in each source,
use \code{test_frame()}.
}
\examples{
\dontrun{
test_register_src("sqlite", {
  DBI::dbConnect(RSQLite::SQLite(), ":memory:", create = TRUE)
})

test_frame(x = 1:3, y = 3:1)
test_load(mtcars)
}
}
\keyword{internal}
