% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-db2.R
\name{backend-db2}
\alias{dialect_db2}
\alias{simulate_db2}
\title{DB2 backend}
\usage{
dialect_db2()

simulate_db2()
}
\description{
This backend supports IBM DB2 databases, typically accessed via ODBC.
Use \code{dialect_db2()} with \code{lazy_frame()} to see simulated SQL without
connecting to a live database.

Key differences for this backend are:
\itemize{
\item Uses \verb{FETCH FIRST n ROWS ONLY} instead of \verb{LIMIT n}
\item Uses double quotes for identifier quoting
}

See \code{vignette("translation-function")} and \code{vignette("translation-verb")} for
details of overall translation technology.
}
\examples{
library(dplyr, warn.conflicts = FALSE)

lf <- lazy_frame(a = TRUE, b = 1, c = 2, d = "z", con = dialect_db2())
lf |> head()
}
