% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-jdbc.R
\name{backend-jdbc}
\title{JDBC backend}
\description{
This backend supports databases accessed via \code{JDBCConnection} created by
\code{RJDBC::JDBC()} and \code{\link[DBI:dbConnect]{DBI::dbConnect()}}. dbplyr automatically detects the
underlying database type from the JDBC connection class and uses the
appropriate SQL dialect.

The following databases are recognized via their JDBC connection class:
\itemize{
\item PostgreSQL (\verb{org.postgresql.*})
\item MySQL (\verb{com.mysql.*})
\item MariaDB (\verb{org.mariadb.*})
\item SQLite (\verb{org.sqlite.*})
\item Oracle (\verb{oracle.*})
\item SQL Server (\verb{com.microsoft.sqlserver.*})
\item SAP HANA (\verb{com.sap.db.*})
\item Teradata (\verb{com.teradata.*})
\item Apache Hive (\verb{org.apache.hive.*})
\item Apache Spark (\verb{org.apache.spark.*} or \verb{com.simba.spark.*})
\item Snowflake (\verb{net.snowflake.*})
\item Impala (\verb{com.cloudera.impala.*})
\item Amazon Redshift (\verb{com.amazon.redshift.*})
}

If your database is not recognized, dbplyr will fall back to a generic ODBC
dialect. Please file an issue if you'd like support for additional databases.

See \code{vignette("translation-function")} and \code{vignette("translation-verb")} for
details of overall translation technology.
}
\examples{
# JDBC connections require the RJDBC package and a JDBC driver JAR file.
# Once connected, dbplyr automatically detects the database type:
#
# library(RJDBC)
# library(dplyr, warn.conflicts = FALSE)
#
# drv <- JDBC("org.postgresql.Driver", "postgresql.jar")
# con <- dbConnect(drv, "jdbc:postgresql://localhost/mydb", "user", "password")
# tbl(con, "my_table") |> filter(x > 1)
}
