% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-odbc.R
\name{backend-odbc}
\alias{dialect_odbc}
\alias{simulate_odbc}
\title{ODBC backend}
\usage{
dialect_odbc()

simulate_odbc()
}
\description{
This backend supports databases accessed via \code{OdbcConnection} created by
\code{\link[DBI:dbConnect]{DBI::dbConnect()}}. Use \code{dialect_odbc()} with \code{lazy_frame()} to see simulated
SQL without connecting to a live database.

Key differences for this backend are minor translations for common data
types.

See \code{vignette("translation-function")} and \code{vignette("translation-verb")} for
details of overall translation technology.
}
\examples{
library(dplyr, warn.conflicts = FALSE)

lf <- lazy_frame(a = TRUE, b = 1, d = 2, c = "z", con = dialect_odbc())
lf |> transmute(x = as.numeric(b))
lf |> transmute(x = as.integer(b))
lf |> transmute(x = as.character(b))
}
