% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind-queries.R
\name{bind_queries}
\alias{bind_queries}
\title{Combine multiple lazy queries}
\usage{
bind_queries(...)
}
\arguments{
\item{...}{<\link[rlang:dyn-dots]{dynamic-dots}> Lazy tables to combine.}
}
\value{
A lazy query.
}
\description{
Combine multiple lazy queries into a single query using \verb{UNION ALL}.
This is a convenient wrapper around \code{purrr::reduce(tables, union_all)}.
}
\examples{
lf1 <- lazy_frame(x = 1, y = "a")
lf2 <- lazy_frame(x = 2, y = "b")
lf3 <- lazy_frame(x = 3, y = "c")

bind_queries(lf1, lf2, lf3)

# If you already have a list, you can use splice operator
queries <- list(lf1, lf2)
bind_queries(!!!queries)
}
