% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db-misc}
\alias{db_desc.DBIConnection}
\alias{sql_join_suffix}
\alias{db_sql_render}
\title{Miscellaneous database generics}
\usage{
\method{db_desc}{DBIConnection}(x)

sql_join_suffix(con, ...)

db_sql_render(con, sql, ...)
}
\description{
Miscellaneous database generics
}
\seealso{
Other generic: 
\code{\link{db-sql}},
\code{\link{db_compute}()},
\code{\link{sql_escape_logical}()}
}
\concept{generic}
