% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db-misc}
\alias{db_desc.DBIConnection}
\alias{sql_join_suffix}
\alias{db_sql_render}
\alias{dbplyr_edition}
\title{Miscellaneous database generics}
\usage{
\method{db_desc}{DBIConnection}(x)

sql_join_suffix(con, ...)

db_sql_render(con, sql, ...)

dbplyr_edition(con)
}
\description{
Use a \code{dbplyr_edition()} method to declare that your backend supports
version 2.0.0 of the dbplyr API where all generics live in dbplyr.
See \code{vignette("backend-2")} for more details.
}
\seealso{
Other generic: 
\code{\link{db-sql}},
\code{\link{db_copy_to}()},
\code{\link{sql_escape_logical}()}
}
\concept{generic}
\keyword{internal}
