% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbplyr.R
\docType{package}
\name{dbplyr-package}
\alias{dbplyr}
\alias{dbplyr-package}
\title{dbplyr: A 'dplyr' Back End for Databases}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

A 'dplyr' back end for databases that allows you to
    work with remote database tables as if they are in-memory data frames.
    Basic features works with any database that has a 'DBI' back end; more
    advanced features require 'SQL' translation to be provided by the
    package author.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dbplyr.tidyverse.org/}
  \item \url{https://github.com/tidyverse/dbplyr}
  \item Report bugs at \url{https://github.com/tidyverse/dbplyr/issues}
}

}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@rstudio.com}

Authors:
\itemize{
  \item Edgar Ruiz
}

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
}

}
\keyword{internal}
