% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verb-do.R
\name{do.tbl_sql}
\alias{do.tbl_sql}
\title{Perform arbitrary computation on remote backend}
\usage{
\method{do}{tbl_sql}(.data, ..., .chunk_size = 10000L)
}
\arguments{
\item{.data}{a tbl}

\item{...}{Expressions to apply to each group. If named, results will be
stored in a new column. If unnamed, must return a data frame. You can
use \code{.} to refer to the current group. You can not mix named and
unnamed arguments.}

\item{.chunk_size}{The size of each chunk to pull into R. If this number is
too big, the process will be slow because R has to allocate and free a lot
of memory. If it's too small, it will be slow, because of the overhead of
talking to the database.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{do()} is deprecated. Instead of \code{do()} you should use \code{\link[dplyr:compute]{dplyr::collect()}} and then
your favourite combination of purrr and dplyr functions.
}
