% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ident.R
\name{ident}
\alias{ident}
\alias{is.ident}
\title{Flag a character vector as SQL identifiers}
\usage{
ident(...)

is.ident(x)
}
\arguments{
\item{...}{A character vector, or name-value pairs.}

\item{x}{An object.}
}
\description{
\code{ident()} takes strings and turns them as database identifiers (e.g. table
or column names) quoting them using the identifier rules for your database.

These are generally for internal use only; if you need to supply an
table name that is qualified with schema or catalog, or has already been
quoted for some other reason, use \code{I()}.
}
\examples{
con <- simulate_dbi()

# SQL92 quotes strings with '
escape("x", con = con)

# And identifiers with "
escape(ident("x"), con = con)
}
\keyword{internal}
