% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ident.R
\name{ident}
\alias{ident}
\alias{is.ident}
\title{Flag a character vector as SQL identifiers}
\usage{
ident(...)

is.ident(x)
}
\arguments{
\item{...}{A character vector, or name-value pairs.}

\item{x}{An object.}
}
\description{
\code{ident()} marks strings as database identifiers (e.g. table or column names)
quoting them using the identifier rules for your database. It is used
primarily in \code{\link[=translate_sql]{translate_sql()}} to label variables as identifiers; use
elsewhere should be regarded with suspicion.

\code{ident()} is for internal use only; if you need to supply an table name that
is qualified with schema or catalog use \code{I()}.
}
\examples{
con <- simulate_dbi()

# SQL92 quotes strings with '
escape("x", con = con)

# And identifiers with "
escape(ident("x"), con = con)
}
\keyword{internal}
