% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ident.R
\name{ident}
\alias{ident}
\alias{is.ident}
\title{Flag a character vector as SQL identifiers}
\usage{
ident(...)

is.ident(x)
}
\arguments{
\item{...}{A character vector, or name-value pairs.}

\item{x}{An object.}
}
\description{
\code{ident()} takes strings and turns them as database identifiers (e.g. table
or column names) quoting them using the identifier rules for your database.
\code{ident_q()} does the same, but assumes the names have already been
quoted, preventing them from being quoted again.

These are generally for internal use only; if you need to supply an
table name that is qualified with schema or catalog, or has already been
quoted for some other reason, use \code{I()}.
}
\examples{
# SQL92 quotes strings with '
escape_ansi("x")

# And identifiers with "
ident("x")
escape_ansi(ident("x"))

# You can supply multiple inputs
ident(a = "x", b = "y")
ident_q(a = "x", b = "y")
}
\keyword{internal}
