% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-sql.R, R/rows.R
\name{sql_returning_cols}
\alias{sql_returning_cols}
\alias{sql_output_cols}
\alias{sql_query_rows_update}
\alias{rows_update.tbl_lazy}
\title{sql_returning_cols}
\usage{
sql_returning_cols(con, cols, table, ...)

sql_output_cols(con, cols, output_delete = FALSE, ...)

sql_query_rows_update(con, x_name, y, by, ..., returning_cols = NULL)

\method{rows_update}{tbl_lazy}(
  x,
  y,
  by = NULL,
  ...,
  copy = FALSE,
  in_place = NULL,
  check = NULL,
  returning = NULL
)
}
\arguments{
\item{output_delete}{For \code{sql_output_cols()}, construct the SQL
for a \code{DELETE} operation.}
}
\description{
\code{sql_returning_cols()} and \code{sql_output_cols()} construct the SQL
required to support the \code{returning} argument.
Two methods are required, because the syntax for SQL Server
(and some other databases) is vastly different from Postgres and other
more standardized DBs.
}
