% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{show_query}
\alias{show_query}
\alias{show_query.tbl_lazy}
\alias{explain.tbl_sql}
\title{Show generated SQL and query plan}
\usage{
\method{show_query}{tbl_lazy}(x, ..., use_colour = TRUE, cte = FALSE, sql_options = NULL)

\method{explain}{tbl_sql}(x, ...)
}
\arguments{
\item{x}{An object to explain}

\item{...}{For \code{explain()}, further arguments to \code{\link[=remote_query_plan]{remote_query_plan()}}.
For \code{show_query()}, ignored.}

\item{use_colour}{Should the output be coloured?}

\item{cte}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Use the \code{sql_options} argument instead.}

\item{sql_options}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
SQL rendering options generated by \code{\link[=sql_options]{sql_options()}}.}
}
\description{
\code{show_sql()} displays the SQL query that will be dispatched to the database;
\code{explain()} displays both the SQL query and the query plan.
}
