% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_dbi}
\alias{simulate_dbi}
\title{Simulate database connections}
\usage{
simulate_dbi(class = character(), ...)
}
\description{
These functions generate S3 objects that have been designed to simulate
the action of a database connection, without actually having the database
available. Obviously, this simulation can only be incomplete, but most
importantly it allows us to simulate SQL generation for any database without
actually connecting to it.

Simulated SQL quotes identifiers with \code{"x"} (double quotes) by default,
\code{`x`} (backticks) for MySQL/MariaDB/SQLite, and \verb{[x]} (square brackets)
for SQL Server. Strings are quoted with \code{'x'}.
}
\keyword{internal}
