% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{sql}
\alias{sql}
\alias{is.sql}
\title{SQL escaping.}
\usage{
sql(...)

is.sql(x)
}
\arguments{
\item{...}{Character vectors that will be combined into a single SQL vector.}

\item{x}{Object to check if it is an sql object.}
}
\description{
These functions are critical when writing functions that translate R
functions to sql functions. Typically a conversion function should escape
all its inputs and return an sql object.
}
