% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{sql}
\alias{sql}
\alias{is.sql}
\alias{as.sql}
\title{SQL escaping.}
\usage{
sql(...)

is.sql(x)

as.sql(x, con, ...)
}
\arguments{
\item{...}{Character vectors that will be combined into a single SQL
expression.}

\item{x}{Object to coerce}

\item{con}{Needed when \code{x} is directly supplied from the user so that
schema specifications can be quoted using the correct identifiers.}
}
\description{
These functions are critical when writing functions that translate R
functions to sql functions. Typically a conversion function should escape
all its inputs and return an sql object.
}
