% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-postgres.R
\name{sql_query_upsert.PqConnection}
\alias{sql_query_upsert.PqConnection}
\title{Upsert using the \verb{ON CONFLICT ... DO UPDATE} clause}
\usage{
\method{sql_query_upsert}{PqConnection}(
  con,
  x_name,
  y,
  by,
  update_cols,
  ...,
  returning_cols = NULL,
  method = NULL
)
}
\description{
The \verb{ON CONFLICT ... DO UPDATE} clause requires a unique index on the columns
in \code{by}. One of the advantages is better concurrency behaviour. See the
\href{https://wiki.postgresql.org/wiki/UPSERT#SQL_MERGE_syntax}{Postgres Wiki} for
more information on the advantages and disadvantages compared to other
upsert implementations.
}
