% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.r, R/tbl-lazy.R
\name{simulate_dbi}
\alias{simulate_dbi}
\alias{simulate_sqlite}
\alias{simulate_postgres}
\alias{simulate_mysql}
\alias{simulate_odbc}
\alias{simulate_impala}
\alias{simulate_mssql}
\alias{tbl_lazy}
\alias{lazy_frame}
\title{Create a local lazy tibble}
\usage{
simulate_dbi()

simulate_sqlite()

simulate_postgres()

simulate_mysql()

simulate_odbc(type = NULL)

simulate_impala()

simulate_mssql()

tbl_lazy(df, src = NULL)

lazy_frame(..., src = NULL)
}
\description{
These functions are useful for testing SQL generation without having to
have an active database connection.
}
\examples{
library(dplyr)
df <- data.frame(x = 1, y = 2)

df_sqlite <- tbl_lazy(df, src = simulate_sqlite())
df_sqlite \%>\% summarise(x = sd(x)) \%>\% show_query()
}
\keyword{internal}
