% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl-lazy.R
\name{tbl_lazy}
\alias{tbl_lazy}
\alias{lazy_frame}
\title{Create a local lazy tibble}
\usage{
tbl_lazy(df, con = NULL, ..., name = "df")

lazy_frame(..., con = NULL, .name = "df")
}
\description{
These functions are useful for testing SQL generation without having to
have an active database connection. See \code{\link[=simulate_dbi]{simulate_dbi()}} for a list
available database simulations.
}
\examples{
library(dplyr)
df <- data.frame(x = 1, y = 2)

df_sqlite <- tbl_lazy(df, con = simulate_sqlite())
df_sqlite \%>\% summarise(x = sd(x, na.rm = TRUE)) \%>\% show_query()
}
\keyword{internal}
