% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl-sql.R
\name{tbl_sql}
\alias{tbl_sql}
\title{Create an SQL tbl (abstract)}
\usage{
tbl_sql(subclass, src, from, ..., vars = NULL, check_from = deprecated())
}
\arguments{
\item{subclass}{name of subclass}

\item{...}{needed for agreement with generic. Not otherwise used.}

\item{vars}{Provide column names as a character vector
to avoid retrieving them from the database.
Mainly useful for better performance when creating
multiple \code{tbl} objects.}

\item{check_from}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
}
\description{
Generally, you should no longer need to provide a custom \code{tbl()}
method.
The default \code{tbl.DBIConnect} method should work in most cases.
}
\keyword{internal}
